Increase default channel window size to a little over 128 kB
30,000 bytes is pretty small.  Even over a 5 ms latency link, this is
insufficient to saturate 100 Mbps.  Increasing to 128 kB will make it
possible to saturate a Gbit link with 1ms latency or a 100 Mbps link
with up to 10 ms latency.

This will also increase the memory used per channel by a factor of 4
from 60,000 bytes to 256 kB.

In practice, I observed the receive window staying at about 16 KiB
(16,384 bytes), so I'm guessing that something in the ChannelManager
logic is rounding it down to a power of two.  Therefore, I set the
default to 144 kB in the hopes that the window will be able to stay at
128 kB, rather than being rounded down to around 64 kB.