Revisiting the logging level.
I noticed that TransportManager.receiveLoop never returns normally and so as it stands now we end up always reporting a warning, even during the normal termination of a connection.

Since trilead is meant to be a general purpose library, it is not desirable for it to produce any logging in the default JVM setting,
which means >=INFO is not OK.

Toning down the whole level by a notch to satisfy this constraint. We can still selectively enable logging levels to get details captured.