[JENKINS-20108]
Prefer SHA1PRNG, which avoids VM-wide global lock on Unix.

Another way to fix this is to avoid repeated calls to "rnd.nextInt()" in
TransportConnection.sendMessage() and instead get all the padding in one call.

That would require another byte array as there's no SecureRandom.nextBytes(byte[],int,int),
but that'd be a monitor cost in the face of VM-wide lock across all the SecureRandom instances.

The amount of padding is fairly small, so yet another way to fix this is to buffer random bytes
by hitting SecureRandom in a bigger block (thus infrequentl) and save the rest at Connection.