Update SFTPv3Client.java, openFile public (#216)
* Update SFTPv3Client.java, openFile public

Current public APIs for open calls are issued with these flags: SSH_FXF_READ|SSH_FXF_WRITE|SSH_FXF_CREAT|SSH_FXF_TRUNC, i.e. for both reading and writing simultaneously. 
due to the way S3 works, this is not feasible to implement it(server) properly, as we would have to store somewhere the whole file being written be able to properly serve the read requests. this could require as much memory or disk volume space as the largest file to be written


I made the existing private method openFile()
public to allow finer control to the few  end users who need it.  


We intend to use it like:
openFile(fileName, 0x00000018 | 0x00000002, attr)
which is the same as existing createFileTruncate but without SSH_FXF_READ

* Update SFTPv3Client.java javadoc fix